+++ tests/UnitTestEx.NUnit.Test/Other/ObjectComparerTest.cs
@@ -1,9 +1,11 @@
 ﻿using NUnit.Framework;
+using System;
 using UnitTestEx.NUnit.Test.Model;
 
 namespace UnitTestEx.NUnit.Test.Other
 {
     [TestFixture]
+    [Obsolete($"This is being replaced by the {nameof(UnitTestEx.Abstractions.JsonElementComparer)} and usage of paths to ignore (versus members) to be more explicit.")]
     public class ObjectComparerTest
     {
         [Test]